third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/watch-expressions-panel-switch.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script src="../../../http/tests/inspector/network-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    var x = Math.sqrt(16);
    debugger;
    return x;
}

function test()
{
    InspectorTest.setQuiet(true);
    InspectorTest.startDebuggerTest(step1);
    var updateCount = 8;

    WebInspector.settings.createLocalSetting("watchExpressions", []).set(["x", "y.foo"]);

    function step1()
    {
        watchExpressionsPane = self.runtime.sharedInstance(WebInspector.WatchExpressionsSidebarPane);
        WebInspector.panels.sources._sidebarPaneStack.showView(WebInspector.panels.sources._watchSidebarPane).then(() => {
            InspectorTest.addSniffer(WebInspector.WatchExpression.prototype, "_createWatchExpression", watchExpressionsUpdated, true);
            InspectorTest.evaluateInPage("testFunction()");
        });
    }

    function watchExpressionsUpdated(result, wasThrown)
    {
        InspectorTest.addResult(this._element.deepTextContent());
        if (--updateCount === 0)
            InspectorTest.completeDebuggerTest();
    }
}

</script>
</head>
<body onload="runTest()">
<p>Tests debugger does not fail when stopped while a panel other than scripts was opened. Both valid and invalid expressions are added to watch expressions.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=70718">Bug 70718</a>
</body>
</html>
