third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/watch-expressions-panel-switch-expected.txt

Tests debugger does not fail when stopped while a panel other than scripts was opened. Both valid and invalid expressions are added to watch expressions.

Bug 70718
x: <not available>
y.foo: <not available>
x: <not available>
y.foo: <not available>
x: <not available>
y.foo: <not available>
x: 4
y.foo: <not available>
